Refine helper method for finding the `Ordering` node for a container, if it has one
We're going to need to find the `Ordering` node more often than just in
the conflict/update detection logic, and the current method of finding a
`Vec` of possible nodes is a bit awkward to be using in multiple places.
This makes it so we have a method returning a
`Result<Option<NodeIndex>>`, which is what we were really wanting in the
existing places anyway.