fix(si): support self update through symbolic links on macOS
When the launcher (i.e. the `si` program) is installed from a non-root
user and when they don't have either `$HOME/.local/bin` or `$HOME/bin`
currently in their `$PATH`, the launcher is installed to
`$HOME/.si/bin/si` and a symlink is created at `/usr/local/bin/si` with
a sudo call (that is, the root user owns the symlink as it is a system
directory).

Prior to this change, when a self update was attempted (via calling `si
update --self) under the scenario above, the new binary would attempt to
write itself to a temp file under the *symlink*'s  parent directory of
`/usr/local/bin`. This would cause an permissions error that looks like:

```
Downloading new binary
Replacing '/usr/local/bin/si' with new binary
Error:
   0: io: Permission denied (os error 13) at path "/usr/local/bin/.si.__temp__6DqCUt"
   1: Permission denied (os error 13) at path "/usr/local/bin/.si.__temp__6DqCUt"

Location:
   <unknown>

Backtrace omitted. Run with RUST_BACKTRACE=1 environment variable to display it.
Run with RUST_BACKTRACE=full to include source snippets.
```

Additionally, if the non-root user happened to have write permission on
`/usr/local/bin`, then the new binary would overwrite the symlink and
install itself to `/usr/local/bin/si`, thus eliminating the symlink.

The detection of a symlink and its resolution by calling
`std::fs::read_link` is provided via a fix in the upstream
[self-replace](https://crates.io/crates/self-replace) crate.

References: mitsuhiko/self-replace#18

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>