Initial implementation of ordered child nodes in workspace graph
While the concept of an ordering node has existed in the workspace
graph, and the conflict/update detection logic, there was nothing to
create (and maintain) ordering nodes. It is now possible to create nodes
that have ordered children (not all children of a node must be in the
ordering), and to re-order the children after it has been created.

This does not yet test removing children from the ordering, when the
edge to them is removed, nor does it yet test any of the conflict/update
detection logic around ordering conflicts & updates.