Refactor consumer for use in the gobbler
The consumer domain has been refactored to support processing deliveries
externally rather than within methods. While the producer and
environment domains have been hardened, the consumer domain is brittle.

The gobbler will consume messages on the queue, but it will have to
notify the original producer of those messages when it is done. That
work does not exist in this commit, but the consuming part does.

Signed-off-by: Nick Gerace <nick@systeminit.com>