Test basic conflict & update detection with ordered containers
In addition to adding new tests for basic conflict & update detection,
this fixes a bug where the "seen" vector clock was being updated when
changing the order of an `OrderingNodeWeight`, instead of updating the
"write" vector clock.

This also removes an unnecessary additional update of the "write" vector
clock of the `EdgeWeight` when adding an edge. By assuming that the
vector clock is already up to date, testing becomes easier & more
stable.