Add gobbler client/server architecture
This commit adds the gobbler client/server architecture. Primarily, it
operates on three crates:

- si-rabbitmq (modified): the crate for interacting with RabbitMQ that
  SI services will use or are using
- gobbler-server (modified): the crate for running an outer management
  consumer loop as well as inner core consumer loops
- gobbler-client (new): the crate for interfacing with the
  gobbler-server and testing the communication between the two

The addition of gobbler-client is at the center of this commit.
Essentially, it is not only being used as the interface to
gobbler-server, but it is also being used as the entrypoint for testing
the gobbler-server lifecycle.

The outer and inner consumer loops provide the ability to manage gobbler
singletons as well as run the gobbler singletons.
The outer loop consumers on the "gobbler-management" stream and its
lifecycle is owned by the gobbler-server. The inner loops are created on
a per-change-set basis and are only spun up and down by requests from
gobbler-clients. The inner loops provide the core work for the
"gobbler". However, in this commit, all the inner loops do is directly
return the message they were sent. This is done for testing purposes to
ensure that all the scaffolding is there and sound.

As far as the inner RabbitMQ communications go, this commit exposes
"OffsetSpecification" for consumers. It is unclear if this will be a
temporary change or not. We should know the exact offset of type u64 in
the future, but for now, we are using what works for testing purposes.
For context, these changes were driven by tests in the gobbler-client library.

As a bonus, this commit also adds si-rabbitmq testin to protect against
regressions. It's minimal for now and may remain that way.

Signed-off-by: Nick Gerace <nick@systeminit.com>