feat(si-pkg,object-tree): add optional kind key to SiPkg
Adds a new key to the root node of an SiPkg that expresses the "kind" of
a package (currently `SiPkgKind::Module` and `SiPkgKind::WorkspaceBackup`).
The "kind" key is optional, and packages without (like the existing
modules), default to `SiPkgKind::Module`. To support this, a new reader
function was added to `object-tree` to support reading a key that may or
may not be in the expected position (`read_key_line_value_opt`).