Fix naming of variable & debug text in ContentNodeWeight
The "kind" was changed to also include the content hash a while back,
and the combined kind + hash is what we're calling a `ContentAddress`.
This renames the vaiable to reflect this refactoring.

The `Debug` implementation for `ContentNodeWeight` was copied from
before `NodeWeight` was refactored into being a wrapper enum around
inner types, so now we're using the correct name for the struct in the
debug output.