fix(si): Infer the container engine based on args being passed
Currently, if a user passes Podman as the engine and a docker-sock
then the CLI will error, We can make an assumption of the engine
based on the param passed and then let the user know we are doing so

```
si --podman-sock /var/run/podman.sock status
WARNING: A `podman_sock` location parameter was passed but no Podman engine was chosen. Continuing with the assumption you meant to choose Podman as the container engine...

Checking for user supplied docker.sock
System Initiative Launcher is in "local" mode
```

```
SI_CONTAINER_ENGINE=podman si --docker-sock /var/run/docker.sock status
WARNING: A `docker_sock` location parameter was passed but no Docker engine was chosen. Continuing with the assumption you meant to choose Docker as the container engine...

Checking for user supplied docker.sock
System Initiative Launcher is in "local" mode
```