feat(dal,web): Remove func name uniqueness, graceful builtin backups
Remove the func name uniqueness condition to ease the import of updated
builtins. A new schema variant will be imported with the same name, and
the connected functions will be recreated, with the original functions
left intact after import.

If a schema variant is net-new (whether a changed builtin or a new
schema variant), it might reference a function from head. This lead to
those functions being skipped from the package and missing on import,
since they were not in the changeset that contained the new schema
variant. This will grab them and create them new, with an identical
name.

Adds a "builtin" boolean to the SchemaVariantView so that we can use
that metadata to figure out which schema variant we want to add to a
module export.

CAVEAT: This makes it possible for a user to create a function with the
same name as an intrinsic (e.g., si:identity), which is likely to cause
problems (although we will skip those on module/workspaceBackup
imports). We still reference these functions by name in a few places, so
a custom function with the same name as an intrinsic that comes first in
the query order would break in the places where we expect the intrinsic.