Ensure buck2 vsix hyperlink is clear
Currently, the buck2 vsix hyperlink in the new BUCK2 doc implies that it
goes to "VS Code", the editor. Now, it should imply that it links to the
extension that can be locally built.

This commit also fixes the Discord badge in the same file to link to the
server.

Signed-off-by: Nick Gerace <nick@systeminit.com>