fix(sdf): per builtin transaction, no dependent update for builtins
Two changes:
  1. Install each builtin on its own Postgres connection so they run
     truly concurrently.
  2. Skip dependent values updates on builtin installation for
     near-instant builtin creation.

I can't figure out a reason to execute dependent value update jobs on
builtin creation. If there is a value that needs to exist on an
attribute for a schema variant in the default context, we can
pre-execute the function during authoring to produce a default value for
that attribute, or explicitly set the default value.  Otherwise, every
value computed by a function will be properly computed when the
component is created, and this is the context where we care about
computing the most up to date values in any case.

This makes builtin installation take about 30 seconds, which is a huge
win to me versus any desire we might have to execute the dependent value
update jobs for the schema variants themselves, before a component is
created.