Add si-rabbitmq DockerHub image
Add si-rabbitmq DockerHub image for ENG-1703 ("Mostly Everything is a
Node or an Edge") graph work. This directory is being added to main to
get the image building and published in advance to the long running dev
branch getting merged.

The new image is not used for anything other than image building,
promoting and publishing.

Signed-off-by: Nick Gerace <nick@systeminit.com>