feat(dal,si-pkg,sdf): Export/import functions on direct children of root
This allows us to define a function on a direct child of root (domain,
resource_value, secrets, etc.) and then have it preserved in module
import/export.

To preserve existing behavior, if a function is not defined on
resource_value, we will automatically attach si:resourcePayloadToValue
to set that value. Otherwise the custom function will be used.

That function must be detached before a custom function is attached to
an asset's resource_value prop.