feat(dal): Create base Component Attribute Values at Component creation
Rather than relying on the dependent values update process to create all
of the data slots that might need to exist, we now create them when
creating a Component.

The previous behavior was necessary when we had multiple levels of
"overlay" possible for data. From least, to most specific: `Schema
Variant`, `Component`, `System`, `Environment`. We have dropped the
levels past `Component` since this initial design, which means that we
only have the `Schema Variant`, and `Component` levels to deal with, and
no longer have to deal with multiple levels that conceptually exist
within the component itself (`Component`, `System`, and `Environment`).
We now know all possible places that the data "might" exist for a given
Component as soon as the Component itself has been created.

So, now when we create a Component, we also create the appropriate
`Attribute Value`s for the component:
 * All `External Provider`
 * All `Internal Provider` (both explicit & implicit)
 * For every `Prop` starting at the root, up to and including the first
   map/array we encounter.
 * For every map we encountered in the previous list, if there is an
   `Attribute Prototype` for a specific key of the map, we also create
   an `Attribute Value` for that specific map entry. This should only be
   happening for things like codegen & qualifications.

Previously, when we were changing component type, we were removing
`Attribute Value`s, and relying on the dependent values update process
to re-create them as necessary. We were also leaving behind the edge of
connected components (and the attribute prototype arguments), which
caused weird front-end behavior, and potentially caused the dependent
values update process to unnecessarily re-create `Attribute Value`s.

Instead, we now leave the `Attribute Value` in place (as dependent
values update will no longer be recreating them), and if you change the
component type to a configuration frame, we remove the edge and trigger
a dependent values update.

Co-Authored-By: Brit Myers <brit@systeminit.com>
Co-Authored-By: Paulo Cabral <paulo@systeminit.com>