chore(buck2): account for prelude files called `BUCK.v2`
I'm not 100% sure why this change was made in the upstream prelude as
the commit was an automated internal Meta repository sync operation.
However it seems wise (and necessary in this changes) to also check for
any `BUCK.v2` files preferentially in addition to `BUCK` files.

References: https://github.com/facebook/buck2-prelude/pull/16

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>