WIP: impl and use set_color for SchemaVariant
Note: postcard won't deserialize into a serde_json::Value so we're
serializing to string first since we need StaticArgumentValue to be
completely agnostic of its type.

Begins a process of changing workspace graph interfaces to accept `impl
Into<Ulid>` instead of `Ulid` so we don't have to litter the code with
`into()` calls.