fix(deps): remove old/duplicated version of unicase
This is going to require a little tracking down to determine why
Reindeer (presumably) is deciding to pull in 2 dependent versions of
`unicase` for the `mime_guess-2.0.4` crate. Specifically, the
`Cargo.lock` has `unicase-2.7.0`, but the resulting `BUCK` rule has
both `:unicase-2.6.0` and `:unicase-2.7.0` as dependencies.

To see this problem before the fix, you can run:

```
❯ buck2 uquery 'rdeps("root//third-party/rust:mime_guess-2.0.4", "root//...")'
Error evaluating expression:
    rdeps("root//third-party/rust:mime_guess-2.0.4", "root//...")
    ^-----------------------------------------------------------^

Caused by:
    0: Error traversing children of root//third-party/rust:mime_guess-2.0.4
    1: Unknown target `unicase-2.6.0` from package `root//third-party/rust`.
       Did you mean one of the 1274 targets in root//third-party/rust:BUCK?
       Maybe you meant one of these similar targets?
         root//third-party/rust:unicase-2.7.0
```

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>