Wire up "Apply Changes" on the frontend to SDF
This restores (some of) the `apply_change_set` API endpoint. It does not
restore any functionality around Actions. An appropriate rebase request
is created to update the base change set of the change set specified in
the API request, and the change set is marked as applied, if the base
change set is updated without any errors, or conflicts.

This does not attempt to report any conflicts (planned for future work).