Fix RabbitMQ infinite hangs when it isn't ready
Fix RabbitMQ infinite hangs by using a timeout-and-retry technique. When
RabbitMQ is booting up and a user tries to create an "Environment" using
the streams port, there can be an infinite hang. That is an issue in its
own right, but this change seeks to avoid it entirely.

On macOS aarch64 with an M1 Max, it appears that RabbitMQ needs about
~3-4 seconds for the streams feature to become available. It is unclear
if other RabbitMQ services need the same amount of time.

Signed-off-by: Nick Gerace <nick@systeminit.com>