fix(graph): touch ordering nodes after children in DFS post-order
This adds a new edge of kind `Ordinal` from an `Ordering` node to the
things it orders, ensuring that ordered nodes are added to the graph
before the Ordering node is touched in the post-order DFS in
`import_subgraph`. These `Ordinal` edges are added in `add_ordered_edge`
and have to be cleaned up in `remove_edge` as well.

Other changes:

  - `name` is now part of the `si` tree.

  - The content store is graceful about receiving the same data from
    concurrent writes (the same value might be produced by two
    concurrent writers, and stored in the content store's in-memory
    cache before being written out). The Content-store is
    content-addressable, so if a value already exists for the hash, we
    should just do nothing on write.