Bulk load more AttributeValue information in PropertyEditorValues summary generation
Rather than calculating "which func is the one that _actually_ generated
this value" on an AttributeValue by AttributeValue basis, we now
generate the base information for all AttributeValues in the Component.

We also stop walking from the AttributeValue to the root if we find a
func that isn't one of the intrinsics, as there is no need to look more
closely towards the root. You can't have dynamic functions that populate
things underneath another dynamic function.