chore(cyclone): refactor `Server` to not require generics externally
This change is pretty exciting (for me, hey I've been working with this
code for a while now) as it delivers a `cyclone_server::Server` which
has no external generic types while still supporting the various
incoming streams internally.

The key insight was to create a `Runnable` trait which is implemented by
an `InnerServer` which has all the trait-bound generics, but stored in a
`Box` making an inner field of `inner: Box<dyn Runnable + Send>`.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>