Calculate AttributeValue dependency graph ahead of time for DependentValuesUpdate
Calculating the dependency graph whenever we need to figure out where
all the data should flow to during a DependentValuesUpdate takes a
significant amount of time when we do the entire calculation from
scratch.

In order to speed up how longs the DependentValuesUpdate takes, we're
now spreading out the cost by calculating the 1-hop "sources" of data
for every AttributeValue, whenever a the AttributeValue, its
AttributePrototype, or its AttributePrototypeArguments are updated. We
then use these pre-calculated 1-hop sources to build the full graph.