fix(sdf): make connect_component_sockets_to_frame_inner iterative
This change converts the `connect_component_sockets_to_frame_inner`
function from an async/recursive strategy to an iterative strategy.
Prior to this change, when working with a large workspace containing
multiple levels of nested frames (like 5 and more), calls to this
function would recurse enough times that it would exhaust the stack
space and cause the program to panic and abort. The iterative approach
guards against an *arbitrarily* deep set of nested frames.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>