chore(telemetry): wait for opentelmetry shutdown with timeout
This change restructures some of the application telemetry setup in an
effort to change the shutdown order of a server/service. Ideally, we'd
like to send graceful shutdown signals to all running tasks and allow
them to clean up and quit. After all outstanding tasks have finished, we
should (in theory) have wrapped up remaining telemetry traces and
events. However, with the batching OpenTelemetry processor, it might
still be in the process of batching and sending data to an OpenTelemetry
collector. In this change, this OpenTelemtry-related subsystem is
shutdown in a dedicated thread with a timeout to guard against any
potential shutdown deadlocking (which was happening when shutting this
subsystem down within a multi-threaded Tokio runtime).

As a result, short run processes such as the Cyclone server have a
chance to submit their tracing data before the process exits, and we can
benefit from Cyclone participating in our multi-service tracing!

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>