Add ability to attach components to frames in the new engine
This is the first pass at adding frames to the new engine. This commit
is primarily focused on restoring the ability to place a new component
or an existing component onto an existing frame in the frontend.

Relevant changes:
- Restore ability to place a new component into a frame in the frontend
- Add "EdgeWeightKind::FrameContains" to replace symbolic edge
  functionality seen on main
- Add ability for Components to determine their parent
- Add integration test for converting a component into a frame and
  attaching a component to that frame

Note: this commit does not guarantee that nested frames, aggregation
frames, and "up" frames will work. It is only concerned with components
within "down" frames.

Signed-off-by: Nick Gerace <nick@systeminit.com>